Double to float

+++ Clustering/src/main/scala/com/examples/DataPoint.scala
@@ -1,6 +1,6 @@
 package com.examples
 
-class DataPoint(x : Double, y : Double) {
+class DataPoint(x : Float, y : Float) {
 	val myX = x
 	val myY = y
 	
@@ -8,19 +8,20 @@ class DataPoint(x : Double, y : Double) {
 		new DataPoint(this.myX + that.myX, this.myY + that.myY)
 	}
 	
-	def /(that: Double) : DataPoint = {
+	def /(that: Float) : DataPoint = {
 		new DataPoint(this.myX / that, this.myY/that)
 	}
 	
-	def EuclideanDistance(other : DataPoint) : Double = {
+	def EuclideanDistance(other : DataPoint) : Float = {
 		val result = Math.sqrt((this.myX - other.myX)*(this.myX - other.myX)
 								+ (this.myY - other.myY)*(this.myY - other.myY))
-		result						
+		result.floatValue()					
 	}
 }
 
 object DataPoint {
+	
 	def random() = {
-		new DataPoint(Math.random(), Math.random())
+		new DataPoint(Math.random().floatValue(), Math.random().floatValue())
 	}
 }
